+++ tests/test_events.py
@@ -13,8 +13,10 @@ import json
 import re
 from pathlib import Path
 
+DEFAULT_XML_FILE = str(Path(__file__).resolve().parents[1] / "LuminariGUI.xml")
+
 class EventSystemTester:
-    def __init__(self, xml_file="../LuminariGUI.xml"):
+    def __init__(self, xml_file=DEFAULT_XML_FILE):
         self.xml_file = xml_file
         self.lua_path = self._find_lua()
         self.test_results = []
@@ -491,7 +493,7 @@ def main():
     import argparse
     
     parser = argparse.ArgumentParser(description='Test event system in LuminariGUI')
-    parser.add_argument('--xml', default='../LuminariGUI.xml', help='XML file to test')
+    parser.add_argument('--xml', default=DEFAULT_XML_FILE, help='XML file to test')
     parser.add_argument('--verbose', '-v', action='store_true', help='Verbose output')
     parser.add_argument('--quiet', '-q', action='store_true', help='Quiet mode')
     parser.add_argument('--list-handlers', action='store_true', help='List event handlers only')