+++ tests/test_system.py
@@ -14,8 +14,10 @@ import re
 import time
 from pathlib import Path
 
+DEFAULT_XML_FILE = str(Path(__file__).resolve().parents[1] / "LuminariGUI.xml")
+
 class SystemTester:
-    def __init__(self, xml_file="../LuminariGUI.xml"):
+    def __init__(self, xml_file=DEFAULT_XML_FILE):
         self.xml_file = xml_file
         self.lua_path = self._find_lua()
         self.test_results = []
@@ -484,7 +486,7 @@ def main():
     import argparse
     
     parser = argparse.ArgumentParser(description='Run system tests for LuminariGUI')
-    parser.add_argument('--xml', default='../LuminariGUI.xml', help='XML file to test')
+    parser.add_argument('--xml', default=DEFAULT_XML_FILE, help='XML file to test')
     parser.add_argument('--verbose', '-v', action='store_true', help='Verbose output')
     parser.add_argument('--quiet', '-q', action='store_true', help='Quiet mode')